Add web API request authentication with JWT

+++ src/main/java/nl/taallijn/has/http/HttpServerVerticle.java
@@ -20,9 +20,14 @@ import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.net.JksOptions;
 import io.vertx.ext.auth.AuthProvider;
+import io.vertx.ext.auth.KeyStoreOptions;
+import io.vertx.ext.auth.User;
+import io.vertx.ext.auth.jwt.JWTAuth;
+import io.vertx.ext.auth.jwt.JWTAuthOptions;
 import io.vertx.ext.auth.shiro.ShiroAuth;
 import io.vertx.ext.auth.shiro.ShiroAuthOptions;
 import io.vertx.ext.auth.shiro.ShiroAuthRealmType;
+import io.vertx.ext.jwt.JWTOptions;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.client.HttpResponse;
@@ -33,6 +38,7 @@ import io.vertx.ext.web.handler.AuthHandler;
 import io.vertx.ext.web.handler.BodyHandler;
 import io.vertx.ext.web.handler.CookieHandler;
 import io.vertx.ext.web.handler.FormLoginHandler;
+import io.vertx.ext.web.handler.JWTAuthHandler;
 import io.vertx.ext.web.handler.RedirectAuthHandler;
 import io.vertx.ext.web.handler.SessionHandler;
 import io.vertx.ext.web.handler.UserSessionHandler;
@@ -89,7 +95,45 @@ public class HttpServerVerticle extends AbstractVerticle {
 			context.response().setStatusCode(302).putHeader("Location", "/").end();
 		});
 
-		Router apiRouter = Router.router(vertx);
+	    Router apiRouter = Router.router(vertx);
+
+	    JWTAuth jwtAuth = JWTAuth.create(vertx, new JWTAuthOptions()
+	      .setKeyStore(new KeyStoreOptions()
+	        .setPath("keystore.jceks")
+	        .setType("jceks")
+	        .setPassword("secret")));
+
+	    apiRouter.route().handler(JWTAuthHandler.create(jwtAuth, "/api/token"));
+
+	    apiRouter.get("/token").handler(context -> {
+	      JsonObject creds = new JsonObject()
+	        .put("username", context.request().getHeader("login"))
+	        .put("password", context.request().getHeader("password"));
+	      auth.authenticate(creds, authResult -> {
+	        if (authResult.succeeded()) {
+	          User user = authResult.result();
+	          user.isAuthorized("create", canCreate -> {
+	            user.isAuthorized("delete", canDelete -> {
+	              user.isAuthorized("update", canUpdate -> {
+	                String token = jwtAuth.generateToken(
+	                  new JsonObject()
+	                    .put("username", context.request().getHeader("login"))
+	                    .put("canCreate", canCreate.succeeded() && canCreate.result())
+	                    .put("canDelete", canDelete.succeeded() && canDelete.result())
+	                    .put("canUpdate", canUpdate.succeeded() && canUpdate.result()),
+	                  new JWTOptions()
+	                    .setSubject("Wiki API")
+	                    .setIssuer("Vert.x"));
+	                context.response().putHeader("Content-Type", "text/plain").end(token);
+	              });
+	            });
+	          });
+	        } else {
+	          context.fail(401);
+	        }
+	      });
+	    });
+		
 		apiRouter.get("/pages").handler(this::apiRoot);
 		apiRouter.get("/pages/:id").handler(this::apiGetPage);
 		apiRouter.post().handler(BodyHandler.create());
@@ -112,10 +156,14 @@ public class HttpServerVerticle extends AbstractVerticle {
 	}
 
 	private void apiDeletePage(RoutingContext context) {
-		int id = Integer.valueOf(context.request().getParam("id"));
-		dbService.deletePage(id, reply -> {
-			handleSimpleDbReply(context, reply);
-		});
+		if (context.user().principal().getBoolean("canDelete", false)) {
+			int id = Integer.valueOf(context.request().getParam("id"));
+			dbService.deletePage(id, reply -> {
+				handleSimpleDbReply(context, reply);
+			});
+		} else {
+			context.fail(401);
+		}
 	}
 
 	private void handleSimpleDbReply(RoutingContext context, AsyncResult<Void> reply) {
@@ -132,33 +180,41 @@ public class HttpServerVerticle extends AbstractVerticle {
 	}
 
 	private void apiUpdatePage(RoutingContext context) {
-		int id = Integer.valueOf(context.request().getParam("id"));
-		JsonObject page = context.getBodyAsJson();
-		if (!validateJsonPageDocument(context, page, "markdown")) {
-			return;
+		if (context.user().principal().getBoolean("canUpdate", false)) {
+			int id = Integer.valueOf(context.request().getParam("id"));
+			JsonObject page = context.getBodyAsJson();
+			if (!validateJsonPageDocument(context, page, "markdown")) {
+				return;
+			}
+			dbService.savePage(id, page.getString("markdown"), reply -> {
+				handleSimpleDbReply(context, reply);
+			});
+		} else {
+			context.fail(401);
 		}
-		dbService.savePage(id, page.getString("markdown"), reply -> {
-			handleSimpleDbReply(context, reply);
-		});
 	}
 
 	private void apiCreatePage(RoutingContext context) {
-		JsonObject page = context.getBodyAsJson();
-		if (!validateJsonPageDocument(context, page, "name", "markdown")) {
-			return;
-		}
-		dbService.createPage(page.getString("name"), page.getString("markdown"), reply -> {
-			if (reply.succeeded()) {
-				context.response().setStatusCode(201);
-				context.response().putHeader("Content-Type", "application/json");
-				context.response().end(new JsonObject().put("success", true).encode());
-			} else {
-				context.response().setStatusCode(500);
-				context.response().putHeader("Content-Type", "application/json");
-				context.response()
-						.end(new JsonObject().put("success", false).put("error", reply.cause().getMessage()).encode());
+		if (context.user().principal().getBoolean("canCreate", false)) {
+			JsonObject page = context.getBodyAsJson();
+			if (!validateJsonPageDocument(context, page, "name", "markdown")) {
+				return;
 			}
-		});
+			dbService.createPage(page.getString("name"), page.getString("markdown"), reply -> {
+				if (reply.succeeded()) {
+					context.response().setStatusCode(201);
+					context.response().putHeader("Content-Type", "application/json");
+					context.response().end(new JsonObject().put("success", true).encode());
+				} else {
+					context.response().setStatusCode(500);
+					context.response().putHeader("Content-Type", "application/json");
+					context.response().end(
+							new JsonObject().put("success", false).put("error", reply.cause().getMessage()).encode());
+				}
+			});
+		} else {
+			context.fail(401);
+		}
 	}
 
 	private boolean validateJsonPageDocument(RoutingContext context, JsonObject page, String... expectedKeys) {
